refactor: move wildcard snapshot materialization
