+++ wildcard_engine.py
@@ -10,8 +10,7 @@ import re
 import threading
 from dataclasses import dataclass
 from pathlib import Path
-from types import MappingProxyType
-from typing import Iterable, Mapping, Sequence
+from typing import Iterable, Sequence
 
 # NumPy is mandatory in supported ComfyUI runtimes and defines the seeded
 # wildcard sampling contract. A stdlib fallback would produce different results.
@@ -55,8 +54,6 @@ try:
         DEFAULT_TEST_WILDCARD_TEXT,
         WILDCARD_DIR_NAME,
         WILDCARD_EXTENSIONS,
-        _WildcardSourceFile,
-        _WildcardSourceState,
         default_wildcard_root,
         ensure_default_wildcard_root,
         parse_wildcard_extra_paths,
@@ -69,12 +66,21 @@ except ImportError:
         DEFAULT_TEST_WILDCARD_TEXT,
         WILDCARD_DIR_NAME,
         WILDCARD_EXTENSIONS,
-        _WildcardSourceFile,
-        _WildcardSourceState,
         default_wildcard_root,
         ensure_default_wildcard_root,
         parse_wildcard_extra_paths,
         resolve_wildcard_roots,
+    )
+
+try:
+    from .easyuse_anima.wildcard.snapshot import (
+        _build_wildcard_snapshot,
+        _WildcardSnapshot,
+    )
+except ImportError:
+    from easyuse_anima.wildcard.snapshot import (
+        _build_wildcard_snapshot,
+        _WildcardSnapshot,
     )
 
 
@@ -133,31 +139,6 @@ WILDCARD_QUANTIFIER_RE = re.compile(
 COUNT_SPEC_RE = re.compile(
     r"(?:(?P<fixed>\d+)|(?P<minimum>\d*)\s*-\s*(?P<maximum>\d*))"
 )
-
-
-@dataclass(frozen=True)
-class _WildcardSnapshot:
-    cache_key: tuple
-    mapping: Mapping[str, tuple[WildcardOption, ...]]
-    wildcard_names: tuple[str, ...]
-    roots: tuple[str, ...]
-    files: tuple[_WildcardSourceFile, ...]
-    cacheable: bool
-
-    def public_signature(self) -> dict:
-        return {
-            "roots": list(self.roots),
-            "files": [
-                {
-                    "root": source.root,
-                    "path": source.relative_path,
-                    "mtime_ns": source.mtime_ns,
-                    "size": source.size,
-                }
-                for source in self.files
-            ],
-        }
-
 
 _SNAPSHOT_CACHE_LIMIT = 16
 _SNAPSHOT_CONDITION = threading.Condition()
@@ -438,33 +419,6 @@ def next_seed(seed, control: str) -> int:
 def has_wildcard_syntax(text: str) -> bool:
     value = str(text or "")
     return bool(DYNAMIC_RE.search(value) or WILDCARD_RE.search(value) or WILDCARD_QUANTIFIER_RE.search(value))
-
-
-def _build_wildcard_snapshot(source_state: _WildcardSourceState) -> _WildcardSnapshot:
-    mapping: dict[str, list[WildcardOption]] = {}
-    cacheable = True
-    for source in source_state.files:
-        root = source_state.roots[source.root_index]
-        try:
-            entries = _wildcard_sources._load_wildcard_file(root, source.path)
-        except OSError:
-            cacheable = False
-            continue
-        for key, options in entries.items():
-            if key not in mapping and options:
-                mapping[key] = options
-
-    frozen_mapping = MappingProxyType(
-        {key: tuple(options) for key, options in mapping.items()}
-    )
-    return _WildcardSnapshot(
-        cache_key=source_state.cache_key,
-        mapping=frozen_mapping,
-        wildcard_names=tuple(sorted(frozen_mapping)),
-        roots=tuple(str(root) for root in source_state.roots),
-        files=source_state.files,
-        cacheable=cacheable,
-    )
 
 
 def _wildcard_snapshot(roots: Iterable[Path]) -> _WildcardSnapshot:
